+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - support external link for reference data
 - disable download setting
 - system and contrast-dark themes
+- close button
 - jwt expire configuration
 
 ## Changed